refactor(lns-cli): move image args (ImageArgs/ImageCommand/…) into the image module
