+++ crates/lns-cli/tests/behaviours/steps/image_cli.rs
@@ -3,8 +3,8 @@ use std::sync::{Arc, Mutex};
 use crate::runner::CliRun;
 use crate::world::BehaviourWorld;
 use cucumber::{given, then, when};
-use lns_cli::cli::ImageArgs;
 use lns_cli::command::parse_args;
+use lns_cli::image::ImageArgs;
 use lns_cli::image::{self, ImageService};
 use lns_cli::integration::LocalBoxFuture;
 use lns_ipc::{ImageInfo, Request, Response};
